backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
sib-api-v3-sdk==7.6.0
flask-cors
python-dotenv==1.0.0
SQLAlchemy==2.0.23